Fixes the sidebar bug
Also hard-codes settings to be the final item in the sidebar in
tpb-clarity-theming. This is the way it was in tpb-clarity-theme and
that seems nice to me.

As of @backstage/core-components 0.9.5 the SidebarContext.Provider was
deprecated and things began using different functions in order to access
the same information. Not sure why it was suddenly a breaking change, I
would've expected them to allow the use of either for a time :/. But
anyways.

https://github.com/backstage/backstage/blob/master/packages/core-components/CHANGELOG.md#095

Signed-off-by: Michael Stergianis <mstergianis@vmware.com>