Aligns tpb-login SignInPageWrapper more closely to DynamicLoginPage
In previous versions of TAP we used a plugin called DynamicLoginPage
which would automatically render cards based on the users app config. It
seems something similar existed in TPB already but with an important
distinction that when no auth providers were configured it would error
as opposed to allowing guest sign in.

https://gitlab.eng.vmware.com/project-star/pstar-backstage-poc/-/blob/develop/plugins/dynamic-login/src/components/DynamicLoginPage/DynamicLoginPage.tsx

Signed-off-by: Michael Stergianis <mstergianis@vmware.com>